Add Arduino 16u2 firmware
This takes the firmware source from the Arduino repository at:

	https://github.com/arduino/Arduino/tree/1c1ecb65ace969ee3897a475969bd81e44710a3e/hardware/arduino/avr/firmwares/atmegaxxu2/arduino-usbserial

The sources are imported verbatim, except for fixing some whitespace
issues. The makefile is stripped down, removing things like debug, coff,
dfu and doxygen, and the MCU type and lufa path are updated. The lufa
library is not present in the Arduino repository, but this adds a link
to the lufa 100807 version using a submodule.